+++ third-party/sfud-1.1.0/ports/fspim/fspim_sfud_core.h
@@ -1,0 +1,31 @@
+/*
+ * Copyright : (C) 2022 Phytium Information Technology, Inc. 
+ * All Rights Reserved.
+ *  
+ * This program is OPEN SOURCE software: you can redistribute it and/or modify it  
+ * under the terms of the Phytium Public License as published by the Phytium Technology Co.,Ltd,  
+ * either version 1.0 of the License, or (at your option) any later version. 
+ *  
+ * This program is distributed in the hope that it will be useful,but WITHOUT ANY WARRANTY;  
+ * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the Phytium Public License for more details. 
+ *  
+ * 
+ * FilePath: fspim_sfud_core.h
+ * Date: 2022-08-19 10:09:31
+ * LastEditTime: 2022-08-19 10:09:31
+ * Description:  This files is for 
+ * 
+ * Modify History: 
+ *  Ver   Who        Date         Changes
+ * ----- ------     --------    --------------------------------------
+ */
+#ifndef FSPIM_SFUD_H
+#define FSPIM_SFUD_H
+
+#include "sfud.h"
+#include "sfud_def.h"
+
+sfud_err FSpimProbe(sfud_flash *flash);
+
+#endif // !